remove scf downloading from package functionality
